agent_service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.75.0",
    "docker>=7.1.0",
    "fastapi>=0.127.0",
    "gitpython>=3.1.45",
    "google-generativeai>=0.8.6",
    "jsonpath-ng>=1.7.0",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "pexpect>=4.9.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tavily-python>=0.7.17",
    "textual>=6.11.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-languages>=1.10.2",
    "uvicorn>=0.40.0",
]
